deps/v8/test/webkit/instance-of-immediates-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This test makes sure that instance of behaves correctly when the value, constructor, or its prototype are immediates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (1 instanceof 1) threw exception TypeError: Expecting a function in instanceof check, but got 1.
PASS ({} instanceof 1) threw exception TypeError: Expecting a function in instanceof check, but got 1.
PASS (obj instanceof 1) threw exception TypeError: Expecting a function in instanceof check, but got 1.
PASS (1 instanceof {}) threw exception TypeError: Expecting a function in instanceof check, but got #<Object>.
PASS ({} instanceof {}) threw exception TypeError: Expecting a function in instanceof check, but got #<Object>.
PASS (obj instanceof {}) threw exception TypeError: Expecting a function in instanceof check, but got #<Object>.
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) is false
PASS (obj instanceof Constructor) is true
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) threw exception TypeError: Function has non-object prototype '1' in instanceof check.
PASS (obj instanceof Constructor) threw exception TypeError: Function has non-object prototype '1' in instanceof check.
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) is false
PASS (obj instanceof Constructor) is false
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) threw exception TypeError: Function has non-object prototype 'null' in instanceof check.
PASS (obj instanceof Constructor) threw exception TypeError: Function has non-object prototype 'null' in instanceof check.
PASS successfullyParsed is true

TEST COMPLETE
